data files updates; various fixes (#668)
* update function WinSet fail hotfix

* added some update debug logging/output

* removed some debug code

* update script error handling improvements

* improved update script error handling

* update script improvements

* more debugging

* fixed poeprices feedback request being repeated

* update script - use full paths for batch commands not part of cmd.exe

* updated data files

* ItemInfo 3.1.01: Maps, Uniques, Divination Cards. Fixes. (#666)

* 3.1.01: Maps, Uniques, Divination Cards. Fixes.

* Updated and rewritten information for maps.
* Updated Uniques and Divination Cards with wiki information.
* NO (!) support for Elder/Shaper mods. Waiting on GGG's advanced tier info for the clipboard, for now.

* PutResultsOnClipboard fix

For issue #663

* Map adjustments

* slight parsing and display changes
* fix for shaped maps

* fixed buyout only option (poe.trade changes)

* fixed unique flask parsing

* hotfix for unique mods having to many values (adv search)

* added full support for unique items with variants (adv search)

* removed debug code

* updated version/notes/fallback

* added simple support for opening an item on poeapp

* data updates/easter eggs

* Updates from ItemInfo (#669)

* updated version

+++ lib/copyUpdate.bat
@@ -1,15 +1,41 @@
 echo off
+
 set updateScriptPath=%1
 set installPath=%2
 set installFolder=%3
 set tempInstallPath=%4
 set tempInstallFolder=%5
+set debug=%6
+set robocopyExe=%SystemRoot%\System32\robocopy.exe
+set findStrExe=%SystemRoot%\System32\findstr.exe
+
+if %%debug EQU 1 (
+	if not defined in_subprocess (cmd /k set in_subprocess=y ^& %0 %*) & exit )
+)
+
+if not exist %updateScriptPath% (
+	dir /b /a %installPath%"\*" | >nul %findStrExe% "^" && (set extractedFilesExist=1) || (set extractedFilesExist=0)
+	if %%extractedFilesExist EQU 0 (
+		set error="Folder with extracted update files is empty."
+		set errorL=20
+	) else (
+		set error="Folder with extracted update files does not exist."
+		set errorL=19
+	)
+	goto EndScript
+)
 
 :: create temporary install folder 
 :: if the installation of new files fails we don't want to have the script deleted completely
-
 if exist %tempInstallPath% (
 	rd /s /q %tempInstallPath%
+	if exist %tempInstallPath% rd /s /q %tempInstallPath% || rem
+	if exist %tempInstallPath% (
+		echo.
+		echo Executed command: rd /s /q %tempInstallPath%
+		echo Failed to remove/clear temporary installation folder. 
+		echo.
+	)
 )
 mkdir %tempInstallPath%
 
@@ -21,7 +47,8 @@ if not exist %tempInstallPath% (
 
 :: copy new script files to temp install directory
 :: https://ss64.com/nt/robocopy-exit.html
-robocopy /s /NFL /NDL /NJH /NJS /nc /np %updateScriptPath% %tempInstallPath%
+echo Executing command: %robocopyExe% %updateScriptPath% %tempInstallPath% /s /NFL /NDL /NJH /NJS /nc /np
+%robocopyExe% %updateScriptPath% %tempInstallPath% /s /NFL /NDL /NJH /NJS /nc /np 
 if %ERRORLEVEL% EQU 16 set errorL=%ERRORLEVEL% & set error="***FATAL ERROR***" & goto EndScript
 if %ERRORLEVEL% EQU 15 set errorL=%ERRORLEVEL% & set error="OKCOPY + FAIL + MISMATCHES + XTRA" & goto EndScript
 if %ERRORLEVEL% EQU 14 set errorL=%ERRORLEVEL% & set error="FAIL + MISMATCHES + XTRA" & goto EndScript
@@ -42,21 +69,52 @@ if %ERRORLEVEL% EQU 0 set errorL=%ERRORLEVEL% & set error="No Change" & goto End
 
 :SwapInstalls
 :: delete install folder (old script) and rename temp install folder to install folder
-rd /s /q %installPath%
-dir /b /a %installPath%"\*" | >nul findstr "^" && (set clearInstallDir=0) || set clearInstallDir=1
+if exist %installPath% (
+	rd /s /q %installPath%
+	if exist %installPath% rd /s /q %installPath% || rem
+	if exist %installPath% (
+		echo.
+		echo executed command: rd /s /q %installPath%
+		echo Failed to remove/clear installation folder. 
+		echo.
+	)
+)
+
+if exist %installPath% (
+	dir /b /a %installPath%"\*" | >nul %findStrExe% "^" && (set clearInstallDir=0) || set clearInstallDir=1
+) else (
+	set clearInstallDir=1
+)
+
 if %clearInstallDir% EQU 1 (
-	robocopy %tempInstallPath% %installPath% /E /MOVE
+	echo Executing command: %robocopyExe% %tempInstallPath% %installPath% /E /MOVE /NFL /NDL /NJH /nc /np
+	%robocopyExe% %tempInstallPath% %installPath% /E /MOVE /NFL /NDL /NJH /nc /np
 	if %ERRORLEVEL% NEQ 1 (
-		set error="Failed to rename temp install folder to install folder. Using temp install folder to run the script."
-		set errorL=17
+		echo Swap directories, rename/move temp folder to install folder:
+		echo. robocopy ERRORLEVEL: %ERRORLEVEL%
+		echo. temp               : %tempInstallPath%
+		echo. install            : %installPath%
+		echo. 
+		set error2="Failed to rename temp install folder to install folder. Using temp install folder to run the script."
+		set errorL2=17
 		goto EndScript
 	) else (
-		rd /s /q %tempInstallPath%
+		set error2="OKCOPY"
+		set errorL2=%ERRORLEVEL%
+		if exist %tempInstallPath% (
+			rd /s /q %tempInstallPath%
+			if exist %tempInstallPath% rd /s /q %tempInstallPath% || rem
+			if exist %tempInstallPath% (
+				echo.
+				echo Executed command: rd /s /q %tempInstallPath%
+				echo Cleanup: Failed to clear temporary installation folder. 
+				echo.
+			)
+		)
 	)
 ) else (
-	echo fail
-	set error="Failed to rename temp install folder to install folder. Using temp install folder to run the script."
-	set errorL=17
+	set error2="Failed to rename temp install folder to install folder. Using temp install folder to run the script."
+	set errorL2=17
 	goto EndScript
 )
 goto EndScript
@@ -66,8 +124,14 @@ goto EndScript
 if exist %updateScriptPath% (
 	for /f %%i in ('rd /s /q %updateScriptPath%') do set test=%%i
 )
-echo %error%
-echo %errorL%
+echo - Copy "update files" to "temp install folder": 
+echo.  - Error/Exit Code: %error%
+echo.  - Errorlevel: %errorL%
+echo. 
+echo - Copy "temp install folder" to "install folder":
+echo.  - Error/Exit Code: %error2%
+echo.  - Errorlevel: %errorL2%
+echo. 
 
 :: write exit code to file
 echo %errorL% >exitCode.txt
